test(pr-advisor): split workflow boundary coverage
